Move wait_for_database_active(...) to shared DBaaS file
