+++ plugins/module_utils/linode_database_shared.py
@@ -2,10 +2,13 @@
 Shared helper functions and structures for all Managed Database modules.
 """
 
-from typing import Any, Callable, Dict, Optional, Set
+from typing import Any, Callable, Dict, Optional, Set, TypeVar
 
+from ansible_collections.linode.cloud.plugins.module_utils.linode_helper import (
+    poll_condition,
+)
 from ansible_specdoc.objects import FieldType, SpecField
-from linode_api4 import ApiError
+from linode_api4 import ApiError, LinodeClient
 
 SPEC_UPDATE_WINDOW = {
     "day_of_week": SpecField(
@@ -86,6 +89,30 @@ SPEC_UPDATE_WINDOW_V2 = {
         description=["The hour to begin maintenance based in UTC time."],
     ),
 }
+
+T = TypeVar("T")
+
+
+def wait_for_database_status(
+    client: LinodeClient,
+    database: T,
+    desired_status: str,
+    timeout: int = 45 * 60,
+    step: int = 2,
+) -> None:
+    """
+    Wait for the given database to enter the given desired status.
+    """
+
+    def __poll_status() -> bool:
+        database._api_get()
+        return database.status == desired_status
+
+    poll_condition(
+        __poll_status,
+        timeout=timeout,
+        step=step,
+    )
 
 
 def validate_allow_list(allow_list: Set[str]) -> None:
